Add an AuthBackend that uses our fledgling SSO system.
This is a proof-of-concept implementation, and it has a number of
issues. Most notably, it doesn't have a way to supply team display
names or user's email addresses. Being able to provide a list of teams
is an accidental feature provided by SSO.

Currently, this allows the user to login via SSO, though logout is
currently broken. Manual logout can be achieved by removing cookies
and session data at both the IDE and SSO server.

I think the way to go here is to create some form of UserInfo strategy
that allows the AuthBackend to concentrate on auth and move the job
of getting user/team information to another place.