Make the module manager load some modules on demand.
During a single request we're mainly going to be interacting with
a single module, so loading all the others is a big waste of effort.
This patch introduces the idea of 'lazy' modules (as separate from
'always' modules) that are only loaded when needed. A small number
of modules remain always loaded.

Change-Id: I07c68d5ca2e69676d72257d000e548c085acd4ac