Fix #1429: Reduce the data returned during LDAP groups lookup.
This attempts to resolve the issues where not all the desired
information was being returned by reducing the amount required.
This is done in two ways - firstly by filtering the groups in LDAP,
rather than entirely in PHP, and secondly by only getting back the
cn, since this is all that's now needed.

Change-Id: I754a131c1b09d0f7dc76045ab4ea0938a625a17d