Fix #1460: Provide better detail using errors from import lint.
While the import lint isn't intended to pick up errors in imported
files (it's only meant to identify them for running pylint against),
it would appear that actually using this data is beneficial.

Instead of just throwing exceptions from loading the imports, we
now see if the error looks useful, and output this in a format that
is then read by the PHP layer, and output to the user. Non-useful
errors to the IDE user are still thrown.

Also add a second test suite that validates this behaviour.

Change-Id: I7da245d8d011978e95a3ad1c9bf9ba8c22cc189d