Implement team status image review.
We now send the browser a bas64 encoded version of the image, which
it can then display to the user, along with the md5sum. The md5 is
the 'value' that gets sent back when the reviewer sets a review
state, and it's this that gets compared to the draft value.

This patch therefore removes the exclusion of images from reviews,
including the rejection of attempts to set their review state, and
the related tests.

This patch does not modify the location of the images once reviewed,
relying instead an external mechanism to do this. The only way the
IDE could do this move is if the web server had write access to the
folder they were eventually served from, which would be a bad thing.

One possible issue that this setup leaves us with is the case where
an image is reviewed & the team uploads a new image before the
reviewed one goes live. This would leave the team without an image
until the new image had also been reviewed (and made live).

Since there are md5 checks all over the place, I don't believe that
the unreviewed image would go live in its place (ie, without review),
however the team would remain in the no-image state for longer than
is ideal.

Change-Id: I735dc9dde69f78fc94a3809fc5e257681fbf5e7e