Fix #1459: Ensure project polling always uses the right team.
It turns out that previously the project polling was using a constant
team value, which was set shortly after the user logged in. For the
teams, this worked fine, as they only had one team. For mentors,
this had the potential for weird bugs, such as this one, as we're
often in more than one team, and switch between them quite a bit.

Change-Id: I5a73e7b84e9a4182fc327fd2ab21009e652954cb