Add support for using external-auth as an auth module.
This adds a new auth module that wraps the 'SSOClient' and its
interface to the external-auth server. Login is functional at this
point, but logout isn't (it will require changes to the client).

I'm not entirely happy with how this is working at the moment, as
the auth module needs to do some direct token munging. A possible
alternative would be to have a dedicated postback page that
duplicates the work done by the auth/authenticate endpoint. This
would, however also need work in the client to allow this to be
specified.

Change-Id: If0aece9e600f17c1861357d7d1830cae8d5e375b