Add support for logout using the external-auth module.
This makes use of the chaneg to the JavaScript handling of the
logout redirection to also log the user out from the external-auth
server (as it expects).

Note: This relies on a revision in the external-auth module that's
not yet generally available. If you want to test this, you'll need
to fetch http://srobo.org/gerrit/external-auth refs/changes/61/461/1
in the external-auth submodule.

Change-Id: Ibcd9491e5128b481d743736edb1b24d287930c4f