Convert all array iterations to use numeric indices.
These could cause problems in IE if attempting to iterate using the
JS for .. in .. structure. While these haven't been specifically
reported as being in error, this is more likely to be as a result
of users not reporting it, or simply not having used the functionality.

In any case, these are all actually arrays, so this should work everywhere.

Change-Id: Ief651228a38dc814a3a974ebe7cc47a9d9955156