Fix #1570: Only fetch the repo once to ensure consistent state.
When we fetch the repo, we clean it up a bit & update it vs its
master repo. Clearly we don't want to be doing this for each of the
files we've been asked to delete, as the cleanup would undo the
deletion we just specified.

TODO: review other usages of the repository() function in this
module for similar issues.

Change-Id: I8060fa9c3f5e369560c24e49b81863d87deaf9be