Avoid locks in the master repo just to clone it.
Expose the checks done to 'ensure' that a given path is a valid
bare repository. This allows this check to be done without needing
to instantiate a GitRepository, removing the need for a lock on it.

This should make per-user clones more efficient, while containing
the same level of safety checks.

Change-Id: I138a2f5ce1ef193b41ccc54b762b2a7760ad2d27