Split the GitRepository class to create a ReadOnly base class.
This allows a switch of read-only usages to using the read-only
base class, and then to have them use shared (read) locks. In turn,
this ought to reduce the number of requests that fail due to lock
conflicts, as the number of possible conflicts will be reduced.

This also switches all usages of the working_path (private) instance
variable to use the workingPath method. This is nicer than making it
protected and better shows that it should be an immutable member.

Change-Id: Ic029eb274b24d7ff3c277e5e967a1d566967a451