Remove all trace of the Java exporter.
It hasn't been used for over a year (was never enabled on
the current server), and no-one seemed to miss it. Given
that even when given the chance most people don't use it,
and that it's caused issues in the past, I'm removing this
as it causes more problems than it solves.

Change-Id: I0b3c29b48a710b67b6633dff83a2531e7e4953d4