Fix #1426: Show the external-clone url in the IDE.
This adds actual functionality to the proj/info endpoint, and a use
of it. This now returns the url that can be used to clone the repo
as its only value. In turn, this value is displayed under the project's
name in the IDE, where any user can clearly see it.

Change-Id: Ib1968408be686f662ecf39de59f78b3e7e50b935