Make the file-list selection persist.
Previously, any update to the tree would clear the selection.
While probably a minor irritant to users, it prevented some
somewhat useful uses of the selection as a display mechanism.

This change means that the selection will persist (for files that
still exist) in the updated tree, when the project hasn't changed.
If the used selects a new team or project, then the selection will
still be cleared, as before.

Change-Id: Ic7dac44b452c8bbffee5890b66102c77af063436