Fix #1007: Add optional by-usage sorting of the projects list.
This further improves the use of Chosen as the selector, by making
it possible to select to have the project list sorted by usage,
rather than alphabetically (though this remains the default).

Note that using the raw usage-order-list like this is a bit rough
around the edges, as it means that a single use of a project takes
a while to filter out, but it's probably good enough for now.

Change-Id: I415455f9a33f7f1feb86377b0b5d110efa3cd39d