Extract error checking core logic into a new model.
This moves the actual handling of the check, and the arrangement
of the results, into a separate class which doesn't have any UI
attached to it. This new class has a collection of unit tests, and
has core IDE functionality injected to allow for this.

Rework the ErrorsPage to wrap this new model, and to provide the
same functionality to external callers. The idea is that eventually
much of the Page's non-UI functionality with be removed out to the
relevant callers, or pushed down into the model as appropriate.
Callers will then interact with the model directly, and the Page
will just be a subscriber that renders the current known errors.

Change-Id: Ia52c9a0f03e37f5b263d2a8e3fc2a29db40d3f5f