Fix #1568: Add support for entering various branded links.
A number of teams are now using Facebook, Twitter & Youtube to post
their prgress, in addition to various blogs. It thus makes sense
for us to support suitably labelled links to them, rather than just
letting the teams put plaintext links into their descriptions.

This relies on an srweb change to be able to show the links.

Change-Id: Ib64cc9a7abf8fb2712bf9864c8c0ee5363cabf07