Fix #1071: Don't bother syntax checking a non-python file.
This is an entirely pointless operation that is actually likely to
result in emitting errors at the user which have no meaning. Prevent
all cases where a syntax check could be run against non-python files
by introducing a common can_check method on the errorspage that is
then used throughout the project and edit pages.

Change-Id: Iba44b4d79152da404093f7a005d07e1bf85af300