Also pass the TLA to the zip builder
In anticipation of the SR2015 kit's zip building script needing the
TLA of the team it's creating the zip for, modify our handling to
do this, and pick up the equivalent change in pyenv.

This relies on pyenv change Ic45dc8b13fcc4f313fed2dea8882d55a67dfa7ba.

Change-Id: Id6e2370f38f92418e13b0339801c0bb9f33f30ea