Fix #2292: Monitor and close all child processes
It turns out that previously we weren't properly closing the streams
for child processes which weren't time-limited (which the PHP docs
say that we should), nor were we properly closing timed-out processes.

Rather than try to wait for the process to close once we've terminated
it, which has no guaruntee of responding quickly, simply queue up
the process for dealing with later and close it when the request
finishes. This should both give the child time to exit and ensure
that we respond to the client in a vaguely timely manner, while
avoiding the zombies we have been seeing.

Add a test which utilises an inline python program which swallows
the initial terminate signal and otherwise inifite loops in order
to simulate a really bad process.

Change-Id: Iad2ab845becfe095a37622ed5e8f477b599d9c0d